requirement.txt
alembic==1.8.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ccvvj6f93d/croots/recipe/certifi_1663615678215/work/certifi
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
Faker==14.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
plotly==5.10.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
tenacity==8.1.0
validators==0.20.0
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
